bz591652 - automatically mark systems as broken if cobbler task fails
If a Cobbler task reports failure during provision, it probably means
there is something wrong with the system that requires manual
intervention. We mark it as "Broken" (so that no jobs are scheduled on
it in the meantime) and send an e-mail to its owner asking them to
investigate.

We only take this action when the Cobbler task reports failure (as
opposed to an error in talking to Cobbler, or any other error condition)
because those other errors might be indicative of a problem elsewhere
e.g. in the lab controller.