hack for missing user_name in TestReportProblem
Trying to fix intermittent failures where the login page does not seem
to have finished loading before we try typing in the user_name field.
sel.open() is supposed to wait for the page to load, but my theory is
that it does not do this in the exception case (which we are explicitly
catching here, for the 403). I have no idea if this hack will help.

Change-Id: I7d56570f6bd84cdadde55b0c814e014685151479