fix for TestLabController: ensure correct distro is picked
When this test is run as part of the full test suite, there will be many
distros with the same name (only install name will be unique). But we
need to make sure the right one is used in this test, otherwise it won't
be present in the lab controller we create and hence the job will
never be scheduled.

Change-Id: I396135fa61750b086e5bac82fe736ccfb486147b