fix up priority attribute on <recipeSet/>
It wasn't included in the RELAX NG schema, and it was also not being
added to cloned recipesets. This patch fixes both those problems, and
adds the attribute to complete-job.xml so that it is covered by tests.

Due to the reference to identity.current.user in bkr/server/jobs.py:306,
we now also need a working CherryPy and TG identity in unit tests. This
patch uses turbogears.testutil to set it up.

Bug: 679678
Change-Id: Icc37e64b70387ee40a02d9877e142a4c745d03fc