introduce a new system_status_duration table
In the same vein as the reservation table, this records start and finish
times for each status that a system has. The table is populated by some
sqlalchemy magic (an AttributeExtension) whenever the system.status
property is set.

The table is needed in order to generate reports on aggregate system
utilisation efficiently.

Bug: 680497
Change-Id: I6f3b402d417051a59454eb373180e67b2e845702