javascript to adjust datetimes to local timezone
Added to master.kid so that it applies to every page. Unfortunately this
also requires adding jquery to every page (by means of
tg.include_widgets, to avoid breaking widgets).

At the moment it is probably Moz-only, since it relies on parsing
datetime strings like:

    new Date('2010/01/01 00:00:00 +00:00')

which appears to be non-standard.

Bug: 658515
Change-Id: I2dc299fc4adaa445d92c669d6ecea68de51b4014