widgets can't be params to other widgets
A (new?) feature of TG widgets is that callable param values will be
called before the widget is rendered. Unfortunately widgets themselves
are callable, so having a widget as a parameter to another widget causes
breakage when TG tries to call the sub-widget at the wrong time.

The fix is to make the widgets be member_widgets, not params (or else
just to stuff them into the params dict in the update_params method,
rather than setting them in the constructor).

Change-Id: I6a7cca4291bfb977881541b569beff9fef1f9dc5