can't access system.id attribute after it is detached
delete_system() calls session.close() which causes the system instance
in the tests to become detached. If we then try to access the system.id
attribute we get an exception like this:

    DetachedInstanceError: Instance <System at 0x50f1110> is not bound
    to a Session; attribute refresh operation cannot proceed

So we need to make sure to grab the system's id before calling
delete_system().

Cherry-picked from 'sqlachemy06' branch.

Bug: 715133
Change-Id: I3d783944fd9c0144001e59521ba8f67f83674215