simplify and fix Distro.all_systems
This patch simplifies the Distro.all_systems method to use only
ORM filter constructs.

This is related to commit e5458d29, which fixed the all_systems method
for the reserveworkflow case, where there is an (unrelated) outer join
on 'arch', but broke other uses of the method. This new version should
work properly in all cases.

Bug: 709883
Change-Id: I92a3982700ce623dff3b9642971d8a37908fef3d