remove distro from all lab controllers when removed from primary mirror
This change allows admins to mark a lab controller as a "primary
mirror", meaning that it is the source from which other lab controllers
sync their distros. When a distro is removed from a primary mirror, it
will be immediately removed from all lab controllers, under the
assumption that the other lab controllers will eventually be sync'ing
the deletion anyway.

Bug: 717424
Change-Id: I1fa20a5c7bcc40ccd4e044ad4fc1d182994c5082