fix /login screwing up NestedVariablesFilter
As part of the fix for bug 663277, we patched TurboGears 1.0.8 so that
when redirecting to /login, the original request's query string is
included in forward_url. So kwargs was always empty.

But the upstream fix in TurboGears 1.1.3 is slightly different -- the
original request parameters are passed as parameters to login, so now we
*are* getting kwargs which reveals that we have essentially the same bug
as TurboGears 1.0.8 did originally.

We need to undo the work of NestedVariablesFilter in /login, so
that the original parameters are preserved correctly.

Change-Id: I6546d5899be257fb35bdc0e8680c610e79b96fda