set MySQL connection character set to utf8
On the Python side we always read/write UTF-8 encoded strings to/from
MySQL, but if we don't explicitly set the connection character set when
connecting the MySQL server may default it to something undesirable like
latin-1. In that case the server would interpret all our UTF-8 strings
as latin-1 bytes and encode them *again* as UTF-8 before storing.

Bug: 749441
Change-Id: Ib97f7da5314e9a952a9a92c928944a79a4061042