Remove system_id from watchdog table
The relationship Watchdog -> System is redundant since Watchdogs are
always associated with a Recipe, and we can grab the System from there.
We can also avoid potential inconsistencies if we don't have to maintain
this relationship separately.

This paves the way for Recipes which do not have an associated System.

Change-Id: Ie5aabf0e8c86c07ee2e7d1faf6bc8b801ba51825