workflow-xslt: Fix jobxml_parsed variable declarations
The jobxml_parsed variable in the JobConfig class was not properly declared
and would cause an AttributeError exception if --job-xml was not provided.

This patch cleans up the usage of the jobxml_parsed variable, and marking it
as a private variable.  It also removes the jobxml_parsed variable from the
JobArgument class where it was never needed or used.

Bug: 840734
Change-Id: I0199e6d4258f6c72f81c65c10075823cd58a7980