allow searching for key with any value in hostRequires XML
Previously a key-value filter in <hostRequires/> no value, such as

    <key_value key="THING" op="==" />

was considered meaningless and was ignored. Now it will find systems
which have that key with *any* value. Note that this is different to

    <key_value key="THING" op="==" value="" />

which only finds systems where that key has the value of empty string.

Similarly, op="!=" with no value will find systems which do not have
that key at all.

Bug: 690063
Change-Id: I72ec2c14d47230a9822f495490f0f791cfa00e70