Fix pxe conflicts between arm and x86 systems.
if filename is not specified for uboot/arm systems then we end up using the
same pxelinux.cfg directory that x86 systems use.  This causes a problem when
a default file is present because the format of that file is not understood
properly by uboot/arm systems.  This is fixed by specifyng a dummy filename
entry which is down a directory level.  All further requests will be from
under that level.