print failures in log-delete to stderr
bkr.server.util.log_to_stream was not replacing the log handler for the
root logger, so messages were still going to the server logs.

This patch also fixes the exit status to be accurate, and adds a new
option --debug to print informational log messages, such as urllib3
connection pool stuff.

Bug: 853849
Change-Id: Ic73e0481aa061847499c74bd6bf3fb7e05d3c1f9