Report "None" for unknown destructive and nda fields in task details
When destructive and nda fields are missing in testinfo.desc, the
value is stored by Beaker as NULL, and bkr task-details reports
destructive=False and nda=False for these tasks. This is dangerous
because it means Beaker is reporting a test as not destructive, or
not under NDA, when it does not know that. Now it reports None
for these fields, and are excluded in xml format output to satisfy
the Relax NG schema.

Bug: 862970
Change-Id: Ie3555b6708b827ecd53b097c43c1132afd71cfd1