Declare 'op' attribute as optional
'op' attribute is declared as mandatory in current beaker-job.rng
for many elements where it is not actually mandatory. Now it's
decalered as an optional attribute. Its default value is == when
it's not given explicitly.

And it's not necessary to declare the references to the name patterns
"op" and "equality_op" as optional in schema now, which means:

        <optional>
                   <ref name="op" />
        </optional>

can just go with:

        <ref name="op">

Bug: 880899
Change-Id: Icb47ee36eea59c01af575ae798898a074f11752f