/distribution/beaker/setup: make /tmp/tito/noarch/ a yum repo when installing git builds
The problem with installing certain RPMs directly from /tmp/tito/noarch
is that it can cause dependency conflicts on the beaker base package if
other subpackages are already installed (for example, upgrading
beaker-server when beaker-lab-controller is already installed). The
simplest thing is to make /tmp/tito/noarch/ a yum repo and
let yum figure everything out.

Change-Id: Ie4b3d798d1ab005d0d6be26d989776abeb7ee7e7