make STABLE really be the default value for --tag in bkr workflow commands
Currently the default distro tag filter in bkr workflow commands is
STABLE, but only when generating job XML (and only if a specific distro
is not named). That means if the workflow command does some other
processing based on the tags given, it may come out inconsistent.

In this particular case, bkr machine-test with no --tag option would
find all the possible osmajor-arch combinations for a system, but then
submit <distro_tag value="STABLE" /> in the resulting job which could
fail if no distros are tagged STABLE.

This patch makes STABLE be treated as the default *everywhere* if no
--tag option is given. (Unfortunately it's a quirk of optparse that we
cannot just pass default=['STABLE'] to the add_option method, because
then any --tag options would be appended to STABLE instead of replacing
it.)

Bug: 876752
Change-Id: I6b6720fc52afa3744cb59d5d187dcef252583679