split machine-hours query into two: one by recipe, one by reservation
The original query was really "recipe-hours", since it was from the
point of view of recipes, which means it includes time spent running on
RHEV but doesn't cover manual reservations. Added a second query for
"machine-hours", from the point of view of system reservations, which
means it includes manual reservations but not time spent running on
RHEV.

It doesn't really make sense to try and combine them, especially given
the difficulties with the arch breakdown on the reservation side (due to
systems supporting multiple arches).

Bug: 877264
Change-Id: Iaaa580899e05fa976e1bd448e6242965e8ac0f53