docs: add missing system content from components.xml
The DocBook version of the Beaker Guide contained a file components.xml,
which was not referenced in the actual book and was mostly duplicate
content. It did have a few useful bits which belong on the Systems page
however.

This content is rescued from pub_doc/Beaker_Guide/en-US/components.xml
in commit bf25a01e.

Change-Id: I256320da751c3bbb0a4d6c4f232abc5683668d1f