restore /start result recording in install_start
This got lost along the way somewhere with commit 662d94ee. It will
change in a future release when we have a Provisioning state, but for
now we will keep compatibility with how it currently works.

Also added a test for install_start, and reshuffled some data setup out
of necessity.

Bug: 591656
Change-Id: I62077daf83b0b2a4d807fa78738bd6e5b118a987