Require CherryPy < 3.0 rather than TurboGears
The __requires__ hack is used in several places to cope with
CherryPy 2 and 3 being installed in parallel. It turns out
using TurboGears as the dependency leads to site-packages being
inserted early on sys.path, so the wrong version of many
modules will be imported if the released bkr RPMs are installed
on a development system.

Depending on CherryPy < 3.0 explicitly rather than implicitly
through TurboGears avoids this problematic behaviour.

Change-Id: Ie6474249d6f03c4eba23095946520ce2ef9ab143