show status as "Updating..." if job is dirty
Now that job status is updated asynchronously it is unavoidable that it
lags behind "reality". So we need some way of indicating to the user
that the status we are showing is not up to date, if the job is marked
dirty. Otherwise it looks like Beaker is broken, for example when
cancelling a job and the user is returned to My Jobs, their job would
appear there as still Running.

Change-Id: Icb4be10beb92a58ba9e57c5159cd78c4963343a2