Number of changes to beaker-sync-tasks
It was found during testing that uploading task RPMs over XML-RPC
is not suitable for the purpose. One reason being, that ``createrepo``
is called for every task upload, which is not very desirable.

The way out is to directly write the task RPMs to the disk and
update the database accordingly. As a consequence, the tool now *have*
to be run on the Beaker server on which the the task RPMs are to be copied.

This patch mainly addresses the changes required to implement the above
mentioned change. The tool does lose a bit of its flexibility (and utility), but
perhaps it should first serve the purpose for which it is written.

Relevant documentation changes as well.

Change-Id: If813d334791736d916eb5d27a5278d6ea5f5190b