remove Fedora17-specific kickstart template, merged into Fedora template
The only difference between Fedora and Fedora17 were some one-off
ARM-specific hacks which were not intended to be kept beyond Fedora 17.
But it's easier to keep those in the main Fedora template surrounded by
conditionals, than maintaining two completely separate copies of the
Fedora templates.

Also updated the tests to test for the ARM-specific hacks, and removed
references to Fedora 16 which is now EOL.

Change-Id: I87becc7d974bba26052aa8f899a689a3f154ed96