Another attempt at fixing this test.
The 'nature' of this test is such that will potentially
fetch a huge number of systems when run as part of the
entire suite.

Hence, this fix limits the systems searched to the ones
loaned to a particular user. I believe it still tests
the same code, so it should still be a good test.