tests: default to hardcoded nonexistent lab controller
The problem with using LabController.query.first() is that, in the
dogfood task, the first lab controller is actually real (it's the lab
controller participating in the tests) and so we end up creating
hundreds of junk power commands which beaker-provision will try to
execute. In itself that isn't harmful, but it makes beaker-provision
fall behind on executing the *real* commands that we want it to run in
the bkr.inttest.labcontroller.test_provision module.

Change-Id: Ifee75ef88126fc6cc8adc1ead866f7f0493bd0f0