docs: top-level index is the new "Beaker help" page
This patch, along with a corresponding patch to beaker-project.org, will
make the top level index of these docs become the replacement for the
old http://beaker-project.org/help.html page.

The end result will be mostly unchanged, which explains some of the
unusual markup which now appears in the index page.

Change-Id: I40e059b1c27273aa0f3798af934e61198c5b98a4