Remove the '/bkr/bkr/tg_widgets/turbogears.widgets/spinnerstopped.png error
This happens because 'server.webpath' is being prepended to the URL
twice. Once via tg.widgets, and then again via tg.url().

This patch overrides the relevant templates with a 'hardcoded' value.
This approach seems to be widely used elsewhere in TG itself, so seems
like a reasonable fix for us to include in Beaker.

Bug: 820775
Change-Id: I30834d2e3193188f3aee81ce0b9d28791cdae257