apply tg.url to system page action variables consistently
All Beaker's system page widgets apply tg.url to the action variable,
but the standard TurboGears form widgets *don't*. So when I converted
the lab info form to a standard form, its action comes out wrong because
tg.url is never applied to it.

This patch adjusts the system page widgets to use action as is, in line
with the standard form widgets, and makes the caller responsible for
calling tg.url on the action before passing it in.

Change-Id: Ia8ec0157368b9a19ef522216204b437252a4f60f