make Flask routing obey TurboGears server.webpath prefix
Currently to make the Beaker application be mounted under a prefix of
/bkr, we do the following in the Apache config:

    WSGIScriptAlias /bkr/ /usr/share/bkr/beaker-server.wsgi/bkr/

and then set server.webpath='/bkr' in the application config to make
TurboGears mount the CherryPy application with the same prefix, so that
everything lines up.

The modern WSGI way of doing this is to rely on SCRIPT_NAME instead, so
the Apache config would be:

    WSGIScriptAlias /bkr /usr/share/bkr/beaker-server.wsgi

and then everything else would (mostly) just work in Flask.
Unfortunately for now we need to prefix all our Flask routes, to make it
match the CherryPy approach.

Change-Id: I81db19a0560d1685b032ea12d4f8a7fcf912e85f