avoid touching ORM internals in lazy_create
I had tried to make lazy_create be clever and handle some type of
relationships automatically, but it proved to be too fragile (depended
on SQLAlchemy internal implementation details which changed in 0.7 and
again in 0.8).

This patch removes the magic and updates callers to pass only columns
instead of relationship properties.

Change-Id: I74e1b787f5fc4d53eddc79693782303e51db748f