fix recipe demarcation on the job page
Removed all the horizontal border noise in the job and recipe tables,
and added a horizontal separator between recipes in a recipe set to
demarcate them more clearly.

Also replaced the "Show Results" pill nav with a button group toggle,
which is slightly more consistent with the rest of the controls on the
page (and with the previous version), and helps demarcate recipes too.

Bug: 1012201
Change-Id: I3e99c969a82ebbffbf0ba3c7c5c3ad7833f2afb4