new EFI boot order snippet
Separated out into its own snippet, to make it easier to override on
a per-system basis if necessary.

Uses a different technique to figure out the new boot order, which does
not rely on BootCurrent (absent on some implementations) and which
preserves all other existing entries in the boot order (to avoid a quirk
in some implementations).

Bug: 1031876
Bug: 1030612
Change-Id: I6d65ac9ab0a9ac03de092863730a0010715c48b4