restore development gunicorn to 8 workers
We need more than 1 worker for certain parts of the web UI to function
properly, and to resemble a production system. For example, the Executed
Tasks widget shows a spinner image but the HTTP request for that will be
blocked on the request for /tasks/do_search so the spinner will never
appear.

Pass --workers 1 to run-server.sh if you want to debug the workers.

Change-Id: Ideca3e2b3fc5d13a4a473f6e34ed15c1f52d8037