"quick info" widgets on system page, nav list instead of tabs
This replaces the (very overgrown) system form with three small "quick
info" boxes. They are designed to show the most important information
about the state of the system as concisely as possible, and to provide
a shortcut for the most common operations that the user might want to do
with the system. Most importantly, the quick info boxes occupy much less
vertical space than the system form, allowing the tabs below (which are
the meat of the page) to be shown more prominently and without scrolling
down.

The horizontal tabs are replaced by a vertical nav list, since there
were more tabs than will fit in a single row.

In subsequent commits the form fields which were in the system form, as
well as all the other functionality that has accreted in there alongside
them, will be moved into new or existing tabs on the system page.

Bug: 1014438
Change-Id: Ifa987590efef6781ddc6dac217e0a35745736ec3