hardware essentials tab on system page
This replaces the remaining hardware-related fields from the system
form.

The distinction between "Essentials" and "Details" is that the essential
fields are needed for running Beaker jobs (such as arch), or cannot be
determined by Beaker's inventory script (such as location), whereas the
details are optional extra hardware information which can be populated
by the inventory script if desired.

Change-Id: I0822cfe40deaeaabbfea512d0bd2c653f4ea9f2e