Add support for renaming systems via CSV import
To enable renaming systems via CSV import, the Systems are exported
along with their ID when the "Systems (for modification)" option is
selected on the export page. The previous functioning of CSV
export (with no system IDs) is available via the "Systems" option.

(Documentation for this change will be added as part of the fix to
Bug:979277)

Bug: 987157
Change-Id: I8ef41524e041a329e282822cd34bad3826165abc