use Backgrid for system activity grid
This avoids the TurboGears DataGrid and @paginate decorator for the
system page, which gives us more freedom (for example having multiple
data grids on the system page, or to port the system page controller to
a Flask handler).

We can also defer loading the activity data until it's viewed for the
first time, which saves request processing time and bandwidth for all
the cases where the user is visiting the system page to do something
other than view activity.

Change-Id: I083a501290f9b4f504b540a7da575775a2d5a1c0