ported Reserve Workflow to Backbone, merged into a single page
Previously the Reserve Workflow was a three-step process: 1. pick distro
tree(s), 2. (optionally) pick system, 3. enter job options and submit.
This patch puts all three steps onto a single page, allowing them to be
completed in any order. In particular, this now lets us link to the
Reserve Workflow with the system pre-filled so that the user can pick
a distro and enter their job options. So we can link to the Reserve
Workflow from the system page, which is the ultimate goal of this patch.

The distro selection part of the widget is also extracted out so that it
can be re-used on the system Provision tab.

Change-Id: If2678f75e09bb42927730153bb6e8252f5b82ead