Provision tab on system page now always provisions directly
Also ported it to Backbone.

This is a second attempt at the final solution for bug 855333. That bug
was ultimately caused by the various ways that the Provision tab changed
its behaviour according to the state of the system.

Now the Provision tab always provisions directly, not through the
scheduler (we can link to the Reserve Workflow for that now). It always
appears regardless of the Automated/Manual status of the system. The tab
will be available whenever the user is allowed to provision the system:
typically, while they have reserved it or if they are its owner (or
granted in its access policy).

Change-Id: I5839e33fb60c6404302d0c21bc9d4c40870a5070