don't strip Extended ASCII control chars in console log
Even though "Extended ASCII" defines characters 128-160 as control
characters, they are used in UTF-8 (and are defined by Windows code
pages and other encodings as well) so we don't want to strip them out in
the console logs. Leaving them in doesn't appear to do any harm, in the
worst case they will be considered an undecodeable character.

Bug: 1054035
Change-Id: I545cfd8b466324a7940b76ed938968300c22f7db