fix order_by for System.command_queue relationship
This was a typo when porting to declarative style. In SQLAlchemy < 0.8
the typo doesn't matter, because SystemActivity.id refers to the
activity.id column, however in SQLAlchemy >= 0.8 it refers to the
system_activity.id column instead, which is wrong in this case.

This also necessitated rearranged the CommandActivity class definition
to be above System.

Change-Id: Ia405307f67a6db926bde7986fe95a00bc45adeda