don't complain about duplicate tag names when editing an existing tag
The UniqueRetentionTag validator was being used when saving an existing
tag, where the tag name has not necessarily changed, so it was
incorrectly reporting a validation error.

However we can't just skip the validation entirely, since we need to
check for duplicate tags *if* the tag name is being changed. So this
patch bypasses the validation and just does the checking by hand with
a flash message on error.

Bug: 1012389
Change-Id: I02e6f69606b25013dd71ddaf440395c9309433be