add Flask to setuptools requires, so that the correct Jinja version is imported
Flask requires Jinja 2.6+ which is provided by a parallel-installed egg
on RHEL6, so Flask needs to be in requires in order to make setuptools
do the necessary path manipulation.

This patch intentionally does not add Flask or Jinja2 to the other
places where we currently require CherryPy < 3, because doing so would
break running from source checkouts due to this setuptools bug:

https://bitbucket.org/pypa/setuptools/issue/6/

This means that under some circumstances running from a source checkout
may still import Jinja 2.2 instead of Jinja 2.6. Currently this is not an
issue since we are not using any Jinja templates with Flask.

Change-Id: I0e83bf725db250539bad0ad1f99328e5eec0befd