clean up copyright/license headers
The Software Freedom Law Centre has a good summary of the issue around
copyright notices and license headers in source code:

http://www.softwarefreedom.org/resources/2012/ManagingCopyrightInformation.html

Copyright notices take a lot of effort to accurately maintain (certainly
the ones we have now are not accurate), and they aren't necessary to
secure copyright anyway. The authoritative record of Beaker's authorship
is always the git repository. Therefore we can omit copyright notices
from all our source files.

The GPL suggests, but does not require, a particular header be placed at
the top of every source file. We can use a shorter version of the notice
which still conveys the intended license for the file while taking up
less space. The complete license is still included in the source tree as
COPYING.

Conveniently, this shorter license header still matches the GPL regexp
in licensecheck which means running licensecheck on a Beaker tarball
will give accurate results.

Change-Id: I0c6914445553788bd243b427ea05aeb0137302a6