tests: attempting to fix timing-sensitivty in test_job_delete
Sometimes the job deletion tests fail because they go to the recipe
page, expecting the recipe to be deleted, but it's not yet. This appears
to be because they are clicking through before the deletion has
completed.

I'm hoping that this patch fixes it by correcting the XPath used to look
for the deleted table row.

Change-Id: I33b7ddd188d572f30cb168f166c5437992201df2