insert commands into the top of the queue instead of appending
System.command_queue is supposed to be in descending order (most recent
command at the top).

Normally the insertion position makes no difference because we will be
querying System.command_queue back in a separate transaction, which
means it will always end up ordered correctly. This change is only
significant for tests which enqueue commands, expecting them to end up
in the right position in the list.

Change-Id: I69ee2c34535fffd662ea40c00a83b7b785da09f1