tests: greenlet failures were not being reported in test_async
If a test function being run inside a greenlet failed, it was printed to
stderr but was never raised in the main context so the failure was
ignored.

Also ensured that all MonitoredSubprocess instances in the tests have
a timeout set. This is a workaround to ensure that there is always at
least one other greenlet alive when we block on the test greenlet.
Gevent is not smart enough to know that we have a SIGCHLD handler which
will spawn new greenlets at a future time, therefore when it sees only
one live greenlet waiting on an event it thinks we have deadlocked. The
proper fix for this would be to use the built-in child watching in
Gevent 1.0 final.

Change-Id: I938aa7303ee7764cd60b91f4893871c8bed6a1d5