tests: need to manually expire osmajor.install_options_by_arch
On newer SQLAlchemy, which is more conservative about only expiring
things from the session when necessary, this test was failing. That's
because the OSMajorInstallOptions.lazy_create call completely bypasses
the ORM session, so SQLAlchemy was not refreshing the
osmajor.install_options_by_arch associationproxy.

The simplest fix is to manually mark it as expired after the lazy_create
call (since really, we are invalidating it behind SQLAlchemy's back).

Change-Id: If2929854fcdc6be70225f10cf848758b77a3fd76