beaker-provision: copy pxelinux.0 into place on startup
This is just as a convenience to cover the most common type of system
(x86-based with BIOS firmware, including x86 VMs).

Updated the docs to indicate that the admin needs to install the right
boot images, with some suggestions on where to get the images. No longer
requiring or recommending Cobbler for this purpose.

Bug: 866765
Change-Id: Ic8d22ef699c70690b82299677d7ca80ed425d0eb