look up packages from recipe directly instead of going through ksmeta
The fact that the recipe packages were passed as a colon-separated
string in ksmeta was just a relic from when that was necessary in
Cobbler. Now we can grab the packages from the recipe object directly,
avoiding the problem with colons being mangled.

However, to avoid breaking compatibility we still need to accept
'packages' is ksmeta. The current behaviour is that it replaces any
packages coming from the recipe, so that behaviour is preserved in this
patch.

Bug: 952635
Change-Id: I487abbb40b757d5490a538ddb71f1594f91d7c12