need to give webassets a writeable cache directory during build
All we do with assets during the package build now is to list the
sources. To do that, webassets *doesn't* actually need or care about the
cache, but it still insists on trying to create it because of
manifest='cache'. So we need to place the cache in a location which the
build will have write access to.

Change-Id: I723aadd6a43739b9881d27c2a0f9a83c584ed7f6