tests: fix OpenStack link test for floating IPs
The floating IP support means that VirtResource.fqdn is always set as
soon as it's created, so the test which was expecting the fqdn to
initially be unset is now invalid.

This should have been with commit
d2aa028b94e40b2f6bd8d897ad50b121b6d63cbd but I didn't notice the failure
because dogfood didn't have a Horizon URL configured yet.

Change-Id: I2c4f1fdfc87db9efe680bbeb4ff1280e65ab9eab