need to include 'bootstrap-select.css' in assets depends
It should have been added to the depends list back when the file was
first added, however the fact that it was missing never really mattered
before.

However now that we are building assets at runtime, and relying on the
webassets environment to find all our source files, the depends list
must be accurate otherwise files will not be copied into the RPM (which
is the problem here).

Change-Id: I3698c6a741a468a734cd92b9a300be12d4724262