new OSMajor.default_install_options method for detected ks meta variables
The first such variables we already have are 'end' and 'systemd'. This
moves the definition of 'end' and 'systemd' to the new method, which now
comes at the very start of the install options inheritance chain.

We also flip the conditional so that by default we assume the new
behaviour, the old behaviour is hardcoded for old distros. That means
new custom distros will work without needing to be hardcoded.

The variables are also now documented, so that admins can override them
if necessary.

Change-Id: I41fffc12fc97770cedab96cd99ced9b564d346da