abort the entire worker process if session.close() fails
If we fail to close the session (for example because of an out-of-memory
condition) we want the entire worker process to abort, because any
subsequent requests will fail due to the transaction left open. Under
mod_wsgi daemon mode the worker will just be replaced with a fresh one.

Also removed superfluous session.rollback() call from the close_session
handler. The commit_or_rollback_session handler would already have tried
to call session.rollback(), so if that failed for some reason then
trying again will only make things worse. session.close() will
ultimately send a ROLLBACK to the database as part of releasing the
connection anyway, which is really what we care about.

Bug: 1120439
Change-Id: Ibd6cd6deb924a9e91f586ff841c4dd24859131ec