workflow-xslt: Fix that incorrect beaker job was generated when using --save-internal-xml
When using --save-internal-xml, the internal XML document needed for
the XSLT processing was deleted, which meant no command line arguments
with variables to the XSLT template would be considered.

This patch ensures that the internal XML document generated will be
saved for future use if --save-internal-xml is used.

It can be argued that the internal XML document could be dumped
after the XSLT processing, which was the case before
commit dd874428d6e90141.  The problem with that is that it makes
debugging XSLT files much harder, if there are an issue which makes
the XSLT processor choke.  Having access to the internal document
in such moments is valueable, as you can do the deeper XSLT
debugging using xsltproc.

Change-Id: If15c0214334fe728444e3038e389c85f742b9f47
Signed-off-by: David Sommerseth <davids@redhat.com>