drop 'systemd' ksmeta variable entirely
In commit 160e4e6b I was intending to keep both 'has_systemd' and
'systemd' for compatibility with Beaker installations where systemd=True
is set for custom distros, or where the systemd variable is used in
custom snippets.

However there is no sane way to obey the 'systemd' variable if it is set
while also correctly populating both 'has_systemd' and 'systemd'.

It was only possible to set 'systemd' since Beaker 0.17.2, it was never
documented, and the main use case (setting systemd=True for custom
distros with systemd) is no longer needed because custom distros are
assumed to have systemd by default. So it should be safe to just drop
the 'systemd' variable entirely.

Change-Id: Ifd5e7b1dabe5f48f97a54519a3529e9f665e4b90