remove hardcoded product names in beaker-import
The individual .treeinfo importers now look at the structure of the
config, rather than the product name, to decide whether they can handle
it or not.

There are actually still hardcoded references to RHEL4 and earlier
(which I don't care enough about to fix) and Fedora (which is difficult
to fix without changing repoids which might be an annoying compatibility
break).

But the key point is that RHEL5-style .treeinfo files (as used by CentOS
and Fedora) can now be imported regardless of the product name.

This patch also adds test coverage for CentOS and RHS, and fixes issues
with the main repo not being imported properly on CentOS and
RHEL7 pre-GA.

Bug: 1070575
Change-Id: Ie60c8728c9c64bc338538cf5d0e3b515a09cd505