tests: fix timing issue with access policy row removal
As of commit be4da184 group names in the access policy widget are
actually links, so [text()=""] no longer works. Unfortunately since it
was a negative XPath using not() the fact that [text()] was no longer
matching did not immediately cause a test failure, so I totally missed
it when merging. It just meant that the test would fail if your browser
is fast enough, because it moves on to refresh the page before the AJAX
request to save the policy has finished.

Change-Id: I690c7d9e9adc3e03ebd7b51c12248b1e1bfa7233